Stream Processor Array: Unset unblockNeeded on short memcpys
Ensure that we don't try to unblock the running thread when it was NOT
actually blocked by a prior memcpy or thread synchronize. This was not only
very rare, but could be masked so long as the running thread signal variable
was found in the blockedThreads map. This patch anticipates the upcoming fix
to the copy engine controller, which makes copies proceed as quickly as the
memory will allow. This controller fix will increase the likelihood of hitting
this bug without this patch.