CudaGPU: Make ClockedObject and Fix Clocking
The CudaGPU used a frequency variable separately from the gem5 clocking
convention, which caused it to fail to pass the clock frequency down to the
shader cores. This patch makes the CudaGPU a gem5 ClockedObject so it hands
the clock down to the shader cores and the LSQs. This smooths out the
runtime curves seen in the coalescing tests.