VI_hammer: Move cache line allocation to data response
Since the VI_hammer GPU L1 caches cannot contain modified data, writebacks are
unnecessary. Because of this, it is unnecessary to allocate space in the data
array when a request misses in the cache, but rather, allocation should happen
when the data is returned from a lower level of the cache hierarchy. This patch
also changes the name of a Writeback_Ack to a Write_Ack to clarify that this
cache is not a writeback cache.