L1 and L2 Controllers: Assign appropriate L2 access latencies
The latency to access the L2 cache from the L1 cache in Fermi is about 120 GPU
cycles, which is mostly domimated by the interconnect between them. Assuming a
dance-hall interconnect, the per-hop interconnect latency is approximately 15
cycles. Adjust the communication latencies between the L1 and L2 to make the
uncontended L2 access latency comparable to Fermi on a per-hop basis.