VI_hammer: Correct L2 select bits calculation and use
When the number of L2s is not a power of 2, the L2 bits calculation rounds down
when it should round up. Correct this by checking for potential rounding error.
Further, the mapping from address to L2 caches used 1 too many bits when
calculating the index of the L2. Subtract 1 from number of L2 select_bits to
fix this. The assert here does not affect simulator performance.