VI_hammer Caches: Update resource stall settings
GPU L1 caches: Since the L1 caches are tightly coupled with the GPU LSQ,
resource stalls can interfere with expected cache behavior. To avoid this
issue, disable the GPU L1 cache resource stalls.

For the GPU L2 cache, we can model throughput restrictions with resource
stalls, link bandwidth limitations or both. Through thorough validation, link
bandwidth restrictions appear to be sufficient to affect appropriate network
back-pressure. Further, it is more reasonable that in GPU memory hierarchies,
each of multiple L2 cache banks should be capable of driving off-chip request
bandwidth to memory controllers at a level that can saturate the channel
bandwidth of a single controller. Using resource stalls to model banking would
require that the network link connecting the L2 cache to memory controllers
be sized appropriately for the number of banks modeled by resource stalls. This
would be messy to implement and would result in strange request contention for
the link.

For now, disable L2 resource stalls until it is clear that L2 cache throughput
restrictions must be added. Note that enabling L2 resource stalls is
parameterized for ease of testing in the future.