ShaderLSQ: Fix clock phase buffer scheduling issues
When calling nextCycle() at any point besides the clock edge of a cycle, it
returns 2 clock edges in the future. Since the LSQ may have a phased clock
compared to the shader, make sure that input buffer and retry handling occurs
on the next clock edge instead of 2 edges in the future.

This change can cause significant performance improvements in cases where a
benchmark has multiple concentrated global memory instructions.