Regressions: Add gem5-style regression tester
To run regression tests similarly to gem5, we add a regressions script that can
be executed from gem5 to compile the specified gem5-gpu binaries and run the
specified tests. This patch adds the regression script, and gem5-gpu build_opts
(which can alternatively simplify the command line to build gem5-gpu).

Note: The regressions script relies on additions to the gem5 patches to include
the regression scripts required on the gem5 sidetests (i.e. in another patch)