VI Hammer: Remove default cache latencies to require setting
The default latencies defined in the GPU L1 and L2 cache are not reasonable,
and they cause performance debugging difficulty if the config files do not set
them. To avoid unset/inappropriate values, remove the defaults to require that
config files specify these latencies.

Note: This problem manifest itself as extremely long latencies for CPU memory
accesses that require a coherence request to go to the GPU L2 cache. This
caused CPU progress in concurrent applications to lag GPU progress
significantly.