Cuda GPU: Fix register device memory bug
When using --access-host-pagetable, the GPU does not need to track device
allocations for translations, because the host will translate the addresses.
Further, if pages are not touched in FS mode, but the library still tries to
pass the allocations to cudaRegisterDeviceMemory, the vtophys call fails. To
fix this, just ignore cudaRegisterDeviceMemory calls when the GPU is using the
host pagetable for translations.