Copy Engine: Robustify configurability
The copy engine is written such that configuring the bandwidth is difficult,
and buffering was unlimited. This patch adds simple, approximate capabilities,
which make it far easier to configure the copy engine. Changes include comments
to give the user an idea of how to configure the copy engine for various PCIe
configurations.

At a later date, parameterization for different PCIe configurations may be
added to other protocol configurations.