ShaderMMU: Schedule pagewalk events
Starting pagewalks instantaneously when needed is somewhat unrealistic, since
hardware would need to allocate a pagewalker and send the translation request
to it. Further, walks should not be started in the same cycle that a
pagewalker completes a walk. Add pagewalk events and schedule them for the
cycle after a pagewalker allocation takes place.

Also, since pagewalk events introduce a walker indices map, use the map to
free available pagewalkers and avoid the loop.