ShaderMMU: Handle fault stack pointer
A CPU thread handling a GPU page fault can trigger numerous spurious callbacks
to the faulting GPU, leaving the ShaderMMU to check if a fault is complete
every time one of the callbacks occurs. This can lead to numerous fault
handling race conditions, corner cases, and unnecessary page walks. To cut down
on these spurious callbacks, track the CPU thread's stack pointer from when it
entered the kernel to handle the GPU page fault. Only callback to the GPU when
the CPU returns to the correct stack pointer location. In the ShaderMMU, add
error checking to validate stack pointer values.