ShaderMMU: Split out function to raise page faults
Remaining bugs with the ShaderMMU involve faults that are raised while the CPU
thread is in kernel mode, but no later ShaderMMU activity actually triggers the
fault to the CPU. This results in GPU deadlock waiting for the page fault that
never gets raised. The first step to address this was to make GPU faults
maskable in the interrupts device, so that they can be raised to the CPU even
if the thread is in kernel mode. Now, we remove the condition that pauses
raising a fault to avoid the deadlock, and thus the 'Pending' interrupt state.
We also split out raising page faults as a separate function which performs
ISA-specific handling and MUST run through to completion. Finally, this patch
adds a timeout event for debugging lost page faults.