ShaderLSQ: Throttle stores out to cache
In real hardware, the store send bandwidth is usually less than 1 cache line
per cycle. Fermi and Maxwell bandwidth was empirically found to be 32B per
cycle, or 1 cache line every 4 cycles. To get the effect of this throttling,
add a check that blocks issue after a store for the number of cycles required
based on the store access size.