VI Hammer: Add missing block deallocations
Some blocks in VI hammer were not getting deallocated when they were set to an
idle state (e.g. by invalidate or other_getx/s messages). While functionally
correct, this caused some bad effects on performance, such as blocks in I in
the L1s getting sent to the L2 upon eviction, in turn evicting valid blocks.
Also, if a valid block was in LRU, that block could be evicted rather than a
block in I. This patch adds in the missing deallocations.

This patch is essentially the same as the patch for MOESI_hammer in gem5